+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat(repositories): added vlc repository
 - feat(repositories): added rpm-fusion repository
 - feat(repositories): added google-chrome repository
+- feat(repositories): added vscode repository
 
 ## 0.0.29 - 2026-08-09